chore: presentation/user 패키지 세분화
