--- src/main/kotlin/io/hhplus/concert/presentation/user/PointRequest.kt
@@ -1,5 +1,0 @@
-package io.hhplus.concert.presentation.user
-
-data class PointRequest(
-    val amount: Long
-)
