--- src/main/kotlin/io/hhplus/concert/presentation/user/PointResponse.kt
@@ -1,5 +1,0 @@
-package io.hhplus.concert.presentation.user
-
-data class PointResponse(
-    val amount: Long
-)
