--- src/main/kotlin/io/hhplus/concert/presentation/user/UserRequest.kt
@@ -1,5 +1,0 @@
-package io.hhplus.concert.presentation.user
-
-data class UserRequest(
-    val name: String
-)
