+++ src/main/kotlin/io/hhplus/concert/presentation/user/request/PointRequest.kt
@@ -1,0 +1,5 @@
+package io.hhplus.concert.presentation.user.request
+
+data class PointRequest(
+    val amount: Long
+)
