+++ src/main/kotlin/io/hhplus/concert/presentation/user/request/UserRequest.kt
@@ -1,0 +1,5 @@
+package io.hhplus.concert.presentation.user.request
+
+data class UserRequest(
+    val name: String
+)
