+++ src/main/kotlin/io/hhplus/concert/presentation/user/response/PointResponse.kt
@@ -1,0 +1,5 @@
+package io.hhplus.concert.presentation.user.response
+
+data class PointResponse(
+    val amount: Long
+)
